# openapi/campushire.openapi.json
{
  "components": {
    "schemas": {
...
      "Body_preview_roster_import_api_v1_institutions__institution_id__roster_imports_preview_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_preview_roster_import_api_v1_institutions__institution_id__roster_imports_preview_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MembershipCreate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "student"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "MembershipCreate",
        "type": "object"
      },
      "MembershipResponse": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "institution_id": {
            "format": "uuid",
            "title": "Institution Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "institution_id",
          "user_id",
          "role",
          "status"
        ],
        "title": "MembershipResponse",
        "type": "object"
      },
      "MembershipStatusUpdate": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "pattern": "^(active|suspended|revoked|graduated)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason"
        ],
        "title": "MembershipStatusUpdate",
        "type": "object"
      },
...
      "UserRole": {
        "enum": [
          "student",
          "tnp_owner",
          "tnp_admin",
          "tnp_reviewer",
          "tnp_auditor"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "CampusHire AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/institutions/{institution_id}/invitations/{invitation_id}/resend": {
      "post": {
        "operationId": "resend_membership_invitation_api_v1_institutions__institution_id__invitations__invitation_id__resend_post",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Resend Membership Invitation Api V1 Institutions  Institution Id  Invitations  Invitation Id  Resend Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend Membership Invitation",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/memberships": {
      "get": {
        "operationId": "read_memberships_api_v1_institutions__institution_id__memberships_get",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MembershipResponse"
                  },
                  "title": "Response Read Memberships Api V1 Institutions  Institution Id  Memberships Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Memberships",
        "tags": [
          "institution memberships"
        ]
      },
      "post": {
        "operationId": "create_membership_api_v1_institutions__institution_id__memberships_post",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Membership",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/memberships/export.csv": {
      "get": {
        "operationId": "export_memberships_api_v1_institutions__institution_id__memberships_export_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Memberships",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/memberships/{membership_id}": {
      "patch": {
        "operationId": "change_membership_status_api_v1_institutions__institution_id__memberships__membership_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Membership Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Membership Status",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/roster-imports": {
      "get": {
        "operationId": "read_roster_imports_api_v1_institutions__institution_id__roster_imports_get",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RosterImportSummary"
                  },
                  "title": "Response Read Roster Imports Api V1 Institutions  Institution Id  Roster Imports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Roster Imports",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/roster-imports/preview": {
      "post": {
        "operationId": "preview_roster_import_api_v1_institutions__institution_id__roster_imports_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_roster_import_api_v1_institutions__institution_id__roster_imports_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Roster Import",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/roster-imports/template": {
      "get": {
        "operationId": "roster_template_api_v1_institutions__institution_id__roster_imports_template_get",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Roster Template",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/roster-imports/{roster_import_id}": {
      "get": {
        "operationId": "read_roster_import_api_v1_institutions__institution_id__roster_imports__roster_import_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "roster_import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Roster Import Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Roster Import",
        "tags": [
          "institution memberships"
        ]
      }
    },
    "/api/v1/institutions/{institution_id}/roster-imports/{roster_import_id}/commit": {
      "post": {
        "operationId": "commit_roster_import_api_v1_institutions__institution_id__roster_imports__roster_import_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "institution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Institution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "roster_import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Roster Import Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Roster Import",
        "tags": [
          "institution memberships"
        ]
      }
    },
...
    "/api/v1/operator/institutions": {
      "post": {
        "operationId": "create_institution_api_v1_operator_institutions_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-operator-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstitutionProvisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionProvisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Institution",
        "tags": [
          "institution provisioning"
        ]
      }
    },
...
  }
}